+++ NEWS.md
@@ -11,6 +11,8 @@ author: "Nelson Gonzabato"
 
 **Other changes**
 
+- `df` was renamed as `old_data` in `multi_model_1`, `newdata` to `new_data`.
+
 - `plot_corr` now directly accepts `data.frame` obejcts. Arguments like `round_values` have also been dropped.
 
 - Fixed DOI to Max Kuhn's paper